refactor: Move epochs/types -> epochs
