+++ app/app_config.go
@@ -52,7 +52,7 @@ import (
 	sudomodulev1 "github.com/NibiruChain/nibiru/v2/api/nibiru/sudo/module"
 	tfmodulev1 "github.com/NibiruChain/nibiru/v2/api/nibiru/tokenfactory/module"
 	devgastypes "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
-	epochstypes "github.com/NibiruChain/nibiru/v2/x/epochs/types"
+	"github.com/NibiruChain/nibiru/v2/x/epochs"
 	evmtypes "github.com/NibiruChain/nibiru/v2/x/evm"
 	"github.com/NibiruChain/nibiru/v2/x/genmsg"
 	inflationtypes "github.com/NibiruChain/nibiru/v2/x/inflation/types"
@@ -75,7 +75,7 @@ var (
 		icatypes.ModuleName,
 
 		evmtypes.ModuleName,
-		epochstypes.ModuleName,
+		epochs.ModuleName,
 		sudo.ModuleName,
 		nutil.TreasuryPoolModuleAccount,
 		wasmtypes.ModuleName,
@@ -96,7 +96,7 @@ var (
 		{Account: icatypes.ModuleName},
 
 		{Account: evmtypes.ModuleName, Permissions: []string{authtypes.Minter, authtypes.Burner}},
-		{Account: epochstypes.ModuleName},
+		{Account: epochs.ModuleName},
 		{Account: sudo.ModuleName},
 		{Account: nutil.TreasuryPoolModuleAccount},
 		{Account: wasmtypes.ModuleName, Permissions: []string{authtypes.Burner}},
@@ -140,7 +140,7 @@ var (
 
 		// --------------------------------------------------------------------
 		// Native x/ Modules
-		epochstypes.ModuleName,
+		epochs.ModuleName,
 		oracletypes.ModuleName,
 		inflationtypes.ModuleName,
 		sudo.ModuleName,
