+++ app/simapp/sim_test.go
@@ -38,7 +38,7 @@ import (
 
 	"github.com/NibiruChain/nibiru/v2/app"
 	devgastypes "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
-	epochstypes "github.com/NibiruChain/nibiru/v2/x/epochs/types"
+	"github.com/NibiruChain/nibiru/v2/x/epochs"
 	inflationtypes "github.com/NibiruChain/nibiru/v2/x/inflation/types"
 	oracletypes "github.com/NibiruChain/nibiru/v2/x/oracle/types"
 	"github.com/NibiruChain/nibiru/v2/x/sudo"
@@ -292,7 +292,7 @@ func TestAppImportExport(t *testing.T) {
 		{oldApp.GetKey(capabilitytypes.StoreKey), newApp.GetKey(capabilitytypes.StoreKey), [][]byte{}},
 		{oldApp.GetKey(authzkeeper.StoreKey), newApp.GetKey(authzkeeper.StoreKey), [][]byte{authzkeeper.GrantKey, authzkeeper.GrantQueuePrefix}},
 		{oldApp.GetKey(devgastypes.StoreKey), newApp.GetKey(devgastypes.StoreKey), [][]byte{}},
-		{oldApp.GetKey(epochstypes.StoreKey), newApp.GetKey(epochstypes.StoreKey), [][]byte{}},
+		{oldApp.GetKey(epochs.StoreKey), newApp.GetKey(epochs.StoreKey), [][]byte{}},
 		{oldApp.GetKey(inflationtypes.StoreKey), newApp.GetKey(inflationtypes.StoreKey), [][]byte{}},
 		{oldApp.GetKey(oracletypes.StoreKey), newApp.GetKey(oracletypes.StoreKey), [][]byte{}},
 		{oldApp.GetKey(sudo.StoreKey), newApp.GetKey(sudo.StoreKey), [][]byte{}},
