+++ app/wasmext/stargate_query.go
@@ -4,7 +4,7 @@ import (
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 
 	devgas "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
-	epochs "github.com/NibiruChain/nibiru/v2/x/epochs/types"
+	epochs "github.com/NibiruChain/nibiru/v2/x/epochs"
 	inflation "github.com/NibiruChain/nibiru/v2/x/inflation/types"
 	"github.com/NibiruChain/nibiru/v2/x/sudo"
 	tokenfactory "github.com/NibiruChain/nibiru/v2/x/tokenfactory/types"
