+++ gosdk/querier.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc"
 
 	devgas "github.com/NibiruChain/nibiru/v2/x/devgas/v1/types"
-	epochs "github.com/NibiruChain/nibiru/v2/x/epochs/types"
+	epochs "github.com/NibiruChain/nibiru/v2/x/epochs"
 	"github.com/NibiruChain/nibiru/v2/x/evm"
 	inflation "github.com/NibiruChain/nibiru/v2/x/inflation/types"
 	tokenfactory "github.com/NibiruChain/nibiru/v2/x/tokenfactory/types"
