+++ proto/nibiru/epochs/v1/genesis.proto
@@ -6,10 +6,9 @@ import "google/protobuf/duration.proto";
 import "google/protobuf/timestamp.proto";
 import "nibiru/epochs/v1/state.proto";
 
-option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs/types";
+option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs";
 
 // GenesisState defines the epochs module's genesis state.
 message GenesisState {
-  repeated nibiru.epochs.v1.EpochInfo epochs = 1
-      [ (gogoproto.nullable) = false ];
+  repeated nibiru.epochs.v1.EpochInfo epochs = 1 [(gogoproto.nullable) = false];
 }
