+++ proto/nibiru/epochs/v1/query.proto
@@ -1,12 +1,12 @@
 syntax = "proto3";
 package nibiru.epochs.v1;
 
+import "cosmos/base/query/v1beta1/pagination.proto";
 import "gogoproto/gogo.proto";
 import "google/api/annotations.proto";
-import "cosmos/base/query/v1beta1/pagination.proto";
 import "nibiru/epochs/v1/state.proto";
 
-option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs/types";
+option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs";
 
 // Query defines the gRPC querier service.
 service Query {
@@ -15,17 +15,19 @@ service Query {
     option (google.api.http).get = "/nibiru/epochs/v1beta1/epochs";
   }
   // CurrentEpoch provide current epoch of specified identifier
-  rpc CurrentEpoch(QueryCurrentEpochRequest)
-      returns (QueryCurrentEpochResponse) {
+  rpc CurrentEpoch(QueryCurrentEpochRequest) returns (QueryCurrentEpochResponse) {
     option (google.api.http).get = "/nibiru/epochs/v1beta1/current_epoch";
   }
 }
 
 message QueryEpochInfosRequest {}
 message QueryEpochInfosResponse {
-  repeated nibiru.epochs.v1.EpochInfo epochs = 1
-      [ (gogoproto.nullable) = false ];
+  repeated nibiru.epochs.v1.EpochInfo epochs = 1 [(gogoproto.nullable) = false];
 }
 
-message QueryCurrentEpochRequest { string identifier = 1; }
-message QueryCurrentEpochResponse { uint64 current_epoch = 1; }
+message QueryCurrentEpochRequest {
+  string identifier = 1;
+}
+message QueryCurrentEpochResponse {
+  uint64 current_epoch = 1;
+}
