+++ proto/nibiru/epochs/v1/state.proto
@@ -5,7 +5,7 @@ import "gogoproto/gogo.proto";
 import "google/protobuf/duration.proto";
 import "google/protobuf/timestamp.proto";
 
-option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs/types";
+option go_package = "github.com/NibiruChain/nibiru/v2/x/epochs";
 
 message EpochInfo {
   // A string identifier for the epoch. e.g. "15min" or "1hour"
