--- x/epochs/client/rest/rest.go
@@ -1,15 +1,0 @@
-package rest
-
-import (
-	"github.com/gorilla/mux"
-
-	"github.com/cosmos/cosmos-sdk/client"
-)
-
-const (
-	MethodGet = "GET"
-)
-
-// RegisterRoutes registers epochs-related REST handlers to a router.
-func RegisterRoutes(clientCtx client.Context, r *mux.Router) {
-}
