+++ x/epochs/epochinfo.go
@@ -1,4 +1,4 @@
-package types
+package epochs
 
 import (
 	"errors"
