+++ x/epochs/epochsmod/genesis.go
@@ -1,0 +1,30 @@
+package epochsmod
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/NibiruChain/nibiru/v2/x/epochs"
+	"github.com/NibiruChain/nibiru/v2/x/epochs/keeper"
+)
+
+// InitGenesis sets epoch info from genesis
+func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState epochs.GenesisState) (err error) {
+	err = genState.Validate()
+	if err != nil {
+		return
+	}
+	for _, epoch := range genState.Epochs {
+		if err = k.AddEpochInfo(ctx, epoch); err != nil {
+			return err
+		}
+	}
+	return
+}
+
+// ExportGenesis returns the capability module's exported genesis.
+func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *epochs.GenesisState {
+	genesis := epochs.DefaultGenesisFromTime(ctx.BlockTime())
+	genesis.Epochs = k.AllEpochInfos(ctx)
+
+	return genesis
+}
