+++ x/epochs/errors.go
@@ -1,4 +1,4 @@
-package types
+package epochs
 
 // DONTCOVER
