+++ x/epochs/export.go
@@ -1,4 +1,4 @@
-package types
+package epochs
 
 import (
 	grpc "google.golang.org/grpc"
