+++ x/epochs/genesis_test.go
@@ -1,126 +1,134 @@
-package epochs_test
+package epochs
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
-
-	"github.com/NibiruChain/nibiru/v2/app"
-	"github.com/NibiruChain/nibiru/v2/x/epochs"
-	"github.com/NibiruChain/nibiru/v2/x/epochs/types"
-	"github.com/NibiruChain/nibiru/v2/x/nutil/testutil/testapp"
 )
 
-func TestEpochsExportGenesis(t *testing.T) {
-	chainStartTime := time.Now().UTC()
+func TestGenesisState_Validate(t *testing.T) {
+	tests := []struct {
+		name      string
+		genState  GenesisState
+		errString string
+	}{
+		{
+			name: "duplicate epochinfo",
+			genState: GenesisState{
+				Epochs: []EpochInfo{
+					{
+						Identifier:              "repeated",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+					{
+						Identifier:              "someOther",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+					{
+						Identifier:              "repeated",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+				},
+			},
+			errString: "epoch identifier should be unique",
+		},
+		{
+			name: "at least one invalid epochinfo",
+			genState: GenesisState{
+				Epochs: []EpochInfo{
+					{
+						Identifier:              "repeated",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+					{
+						Identifier:              "someOther",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+					{
+						Identifier:              "the invalid",
+						StartTime:               time.Now(),
+						Duration:                0,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+				},
+			},
+			errString: "epoch duration should NOT be 0",
+		},
+	}
 
-	app, _ := testapp.NewNibiruTestApp(app.GenesisState{})
-	ctx := testapp.NewContext(app).WithBlockTime(chainStartTime)
-
-	genesis := epochs.ExportGenesis(ctx, *app.EpochsKeeper)
-	require.Len(t, genesis.Epochs, 4)
-
-	errMsg := fmt.Sprintf("app.EpochsKeeper.AllEpochInfos(ctx): %v\n", app.EpochsKeeper.AllEpochInfos(ctx))
-	require.Equal(t, genesis.Epochs[0].Identifier, "30 min")
-	require.WithinDurationf(t, genesis.Epochs[0].StartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[0].Duration, time.Minute*30, errMsg)
-	require.Equal(t, genesis.Epochs[0].CurrentEpoch, uint64(0))
-	require.Equal(t, genesis.Epochs[0].CurrentEpochStartHeight, int64(0))
-	require.WithinDurationf(t, genesis.Epochs[0].CurrentEpochStartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[0].EpochCountingStarted, false)
-
-	require.Equal(t, genesis.Epochs[1].Identifier, "day")
-	require.WithinDurationf(t, genesis.Epochs[1].StartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[1].Duration, time.Hour*24)
-	require.Equal(t, genesis.Epochs[1].CurrentEpoch, uint64(0))
-	require.Equal(t, genesis.Epochs[1].CurrentEpochStartHeight, int64(0))
-	require.WithinDurationf(t, genesis.Epochs[1].CurrentEpochStartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[1].EpochCountingStarted, false)
-
-	require.Equal(t, genesis.Epochs[2].Identifier, "month")
-	require.WithinDurationf(t, genesis.Epochs[2].StartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[2].Duration, time.Hour*24*30)
-	require.Equal(t, genesis.Epochs[2].CurrentEpoch, uint64(0))
-	require.Equal(t, genesis.Epochs[2].CurrentEpochStartHeight, int64(0))
-	require.WithinDurationf(t, genesis.Epochs[2].CurrentEpochStartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[2].EpochCountingStarted, false)
-
-	require.Equal(t, genesis.Epochs[3].Identifier, "week")
-	require.WithinDurationf(t, genesis.Epochs[3].StartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[3].Duration, time.Hour*24*7)
-	require.Equal(t, genesis.Epochs[3].CurrentEpoch, uint64(0))
-	require.Equal(t, genesis.Epochs[3].CurrentEpochStartHeight, int64(0))
-	require.WithinDurationf(t, genesis.Epochs[3].CurrentEpochStartTime, chainStartTime, time.Second, errMsg)
-	require.Equal(t, genesis.Epochs[3].EpochCountingStarted, false)
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			err := tc.genState.Validate()
+			require.ErrorContains(t, err, tc.errString)
+		})
+	}
 }
 
-func TestEpochsInitGenesis(t *testing.T) {
-	app, ctx := testapp.NewNibiruTestAppAndContext()
-	// On init genesis, default epochs information is set
-	// To check init genesis again, should make it fresh status
-	epochInfos := app.EpochsKeeper.AllEpochInfos(ctx)
-	for _, epochInfo := range epochInfos {
-		err := app.EpochsKeeper.DeleteEpochInfo(ctx, epochInfo.Identifier)
-		require.NoError(t, err)
-	}
-
-	now := time.Now()
-	ctx = ctx.WithBlockHeight(1)
-	ctx = ctx.WithBlockTime(now)
-
-	// test genesisState validation
-	genesisState := types.GenesisState{
-		Epochs: []types.EpochInfo{
-			{
-				Identifier:              "monthly",
-				StartTime:               time.Time{},
-				Duration:                time.Hour * 24,
-				CurrentEpoch:            0,
-				CurrentEpochStartHeight: ctx.BlockHeight(),
-				CurrentEpochStartTime:   time.Time{},
-				EpochCountingStarted:    true,
-			},
-			{
-				Identifier:              "monthly",
-				StartTime:               time.Time{},
-				Duration:                time.Hour * 24,
-				CurrentEpoch:            0,
-				CurrentEpochStartHeight: ctx.BlockHeight(),
-				CurrentEpochStartTime:   time.Time{},
-				EpochCountingStarted:    true,
-			},
-		},
-	}
-	err := epochs.InitGenesis(ctx, *app.EpochsKeeper, genesisState)
-	require.Error(t, err)
-
-	require.EqualError(t, genesisState.Validate(), "epoch identifier should be unique")
-
-	genesisState = types.GenesisState{
-		Epochs: []types.EpochInfo{
-			{
-				Identifier:              "monthly",
-				StartTime:               time.Time{},
-				Duration:                time.Hour * 24,
-				CurrentEpoch:            0,
-				CurrentEpochStartHeight: ctx.BlockHeight(),
-				CurrentEpochStartTime:   time.Time{},
-				EpochCountingStarted:    true,
+func TestGenesisState_HappyPath(t *testing.T) {
+	tests := []struct {
+		name     string
+		genState GenesisState
+	}{
+		{
+			name: "duplicate epochinfo",
+			genState: GenesisState{
+				Epochs: []EpochInfo{
+					{
+						Identifier:              "firstOne",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+					{
+						Identifier:              "someOther",
+						StartTime:               time.Now(),
+						Duration:                10,
+						CurrentEpoch:            2,
+						CurrentEpochStartTime:   time.Now(),
+						EpochCountingStarted:    false,
+						CurrentEpochStartHeight: 1,
+					},
+				},
 			},
 		},
 	}
 
-	err = epochs.InitGenesis(ctx, *app.EpochsKeeper, genesisState)
-	require.NoError(t, err)
-	epochInfo, err := app.EpochsKeeper.GetEpochInfo(ctx, "monthly")
-	require.NoError(t, err)
-	require.Equal(t, epochInfo.Identifier, "monthly")
-	require.Equal(t, epochInfo.StartTime.UTC().String(), now.UTC().String())
-	require.Equal(t, epochInfo.Duration, time.Hour*24)
-	require.Equal(t, epochInfo.CurrentEpoch, uint64(0))
-	require.Equal(t, epochInfo.CurrentEpochStartHeight, ctx.BlockHeight())
-	require.Equal(t, epochInfo.CurrentEpochStartTime.UTC().String(), time.Time{}.String())
-	require.Equal(t, epochInfo.EpochCountingStarted, true)
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			err := tc.genState.Validate()
+			require.NoError(t, err)
+		})
+	}
 }
