+++ x/epochs/hooks.go
@@ -1,4 +1,4 @@
-package types
+package epochs
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
