+++ x/epochs/identifier.go
@@ -1,4 +1,4 @@
-package types
+package epochs
 
 import (
 	"fmt"
