+++ x/epochs/keeper/keeper.go
@@ -7,15 +7,15 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/codec"
 
-	"github.com/NibiruChain/nibiru/v2/x/epochs/types"
+	"github.com/NibiruChain/nibiru/v2/x/epochs"
 )
 
 type Keeper struct {
 	cdc      codec.Codec
 	storeKey storetypes.StoreKey
-	hooks    types.EpochHooks
+	hooks    epochs.EpochHooks
 
-	Epochs collections.Map[string, types.EpochInfo]
+	Epochs collections.Map[string, epochs.EpochInfo]
 }
 
 func NewKeeper(cdc codec.Codec, storeKey storetypes.StoreKey) *Keeper {
@@ -23,11 +23,11 @@ func NewKeeper(cdc codec.Codec, storeKey storetypes.StoreKey) *Keeper {
 		cdc:      cdc,
 		storeKey: storeKey,
 
-		Epochs: collections.NewMap[string, types.EpochInfo](storeKey, 1, collections.StringKeyEncoder, collections.ProtoValueEncoder[types.EpochInfo](cdc)),
+		Epochs: collections.NewMap[string, epochs.EpochInfo](storeKey, 1, collections.StringKeyEncoder, collections.ProtoValueEncoder[epochs.EpochInfo](cdc)),
 	}
 }
 
 // SetHooks Set the epoch hooks.
-func (k *Keeper) SetHooks(eh types.EpochHooks) {
+func (k *Keeper) SetHooks(eh epochs.EpochHooks) {
 	k.hooks = eh
 }
